fix formatting files names issue
