+++ src/utils/data-formatters/format-workflow-history-event/format-decision-task-started-event.ts
@@ -19,29 +19,18 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import formatDurationToSeconds from '../format-duration-to-seconds';
-import formatEnum from '../format-enum';
-
 import formatWorkflowCommonEventFields from './format-workflow-common-event-fields';
-import { type DecisionTaskScheduledEvent } from './format-workflow-history-event.type';
+import { type DecisionTaskStartedEvent } from './format-workflow-history-event.type';
 
-const formatDecisionTaskScheduledEvent = ({
-  decisionTaskScheduledEventAttributes: {
-    startToCloseTimeout,
-    taskList,
-    ...eventAttributes
-  },
+const formatDecisionTaskStartedEvent = ({
+  decisionTaskStartedEventAttributes: { scheduledEventId, ...eventAttributes },
   ...eventFields
-}: DecisionTaskScheduledEvent) => {
+}: DecisionTaskStartedEvent) => {
   return {
     ...formatWorkflowCommonEventFields(eventFields),
     ...eventAttributes,
-    taskList: {
-      kind: formatEnum(taskList?.kind, 'TASK_LIST_KIND'),
-      name: taskList?.name || null,
-    },
-    startToCloseTimeoutSeconds: formatDurationToSeconds(startToCloseTimeout),
+    scheduledEventId: parseInt(scheduledEventId),
   };
 };
 
-export default formatDecisionTaskScheduledEvent;
+export default formatDecisionTaskStartedEvent;
